RFC-002:Updated Regression tests to work with monkeypatch edits.
