+++ lib/buffer-binding.js
@@ -36,7 +36,9 @@ class BufferBinding {
       this.pushChange(this.pendingChanges.shift())
     }
     this.pendingChanges = null
-    if(!this.isHost) this.addFile(bufferProxy.uri)
+    if (this.isHost !== undefined && !this.isHost) {
+      this.addFile(bufferProxy.uri)
+    }
     this.bufferDestroySubscription = this.buffer.onDidDestroy(() => {
       if (this.isHost) {
         bufferProxy.dispose()
@@ -164,8 +166,8 @@ class BufferBinding {
     }
   }
 
-  didChangeURI(uri){
-    if(this.bufferFile) this.bufferFile.setURI(uri)
+  didChangeURI (uri) {
+    if (this.bufferFile) this.bufferFile.setURI(uri)
   }
 
   getBufferProxyURI () {
