+++ lib/editor-binding.js
@@ -2,7 +2,6 @@
 
 const path = require('path')
 const {Range, Emitter, Disposable, CompositeDisposable} = require('atom')
-const getPathWithNativeSeparators = require('./get-path-with-native-separators')
 const getEditorURI = require('./get-editor-uri')
 const {FollowState} = require('@atom/teletype-client')
 
@@ -71,7 +70,7 @@ class EditorBinding {
 
     let remoteEditorCountForBuffer = buffer.remoteEditorCount || 0
     buffer.remoteEditorCount = ++remoteEditorCountForBuffer
-    buffer.getPath = () => `${prefix}:${editorProxy.bufferProxy.delegate.bufferFile.getURI()}`
+    buffer.getPath = () => `${prefix}:${this.editorProxy.bufferProxy.delegate.bufferFile.getURI()}`
     buffer.save = () => { bufferProxy.requestSave() }
     buffer.isModified = () => false
