+++ test/helpers/fake-editor-proxy.js
@@ -1,20 +1,18 @@
+const FakeBufferProxy = require('./fake-buffer-proxy')
+const BufferBinding = require('../../lib/buffer-binding')
+const {TextBuffer} = require('atom')
+
 let nextEditorProxyId = 1
 
 module.exports =
 class FakeEditorProxy {
   constructor (uri) {
     this.id = nextEditorProxyId++
-    this.bufferProxy = {
-      uri,
-      dispose () {},
-      setDelegate () {},
-      createCheckpoint () {},
-      groupChangesSinceCheckpoint () {},
-      applyGroupingInterval () {},
-      getHistory () {
-        return {undoStack: [], redoStack: [], nextCheckpointId: 1}
-      }
-    }
+    const buffer = new TextBuffer('test?')
+    const binding = new BufferBinding({buffer, isHost: false})
+    this.bufferProxy = new FakeBufferProxy(binding, buffer.getText())
+    this.bufferProxy.uri = uri
+    binding.setBufferProxy(this.bufferProxy)
   }
 
   dispose () {
